code: heat: add a new application to manage heating devices
This application connects over MQTT and publishes a few values, for
now only the measurement of the light sensor to detect people's presence.

It also subscribes to some topics to get PWM values to be used during day
and night, as well as an optional change in the limit between day and night.

For now it checks these elements every second as it's easier for debugging,
but the frequency could be arranged and a cache could be used for unchanged
values.

The PWM output is not affected yet, only the light is read. No temperature
is read either.

All variables and MQTT information are detailed in the doc.