code: heat: support fast heating when leaving ECO mode
When leaving ECO mode, we switch to FAST state for at most as long as
we were in ECO mode, limited to 1000 seconds. One of the limits is that
if the heater is turned on cold without going to ECO mode, we don't
know how to turn it to FAST mode. If we did it it would cause the heater
to be set to full throttle upon each turn on, which is not desirable.