code: make the onboard led blink to reflect activity and mode
During each call to the timer, we'll blink the onboard led if it is
defined (brd_led, brd_led_inv). In ECO mode, the led is off by default
and flashes on. In other modes, the led is on by default and flashes
off. Note that the onboard LED on ESP12E and newer is connected to
GPIO2 (D4) which is also the pin that was used to connect the DHT
module on some boards.