conf: add a variant for Pilot Wire
This variant connects a MOC3041 triac to the PWM output which is then
used in full-on, full-off mode. It would be interesting to see if by
modulating it at 50 Hz and using a half-wave ratio it's possible to
switch to the -1 or -2 degrees that some devices support, or if this
requires phase synchronization.