schematic: add the schematic, board and corresponding images
This will ease duplicating the device. The circuit is purpusely single
sided, with all the generic stuff under the ESP and the project-specific
one around, so that adapting it to any other use case should be easy.
Particular care was given to make the default circuit entirely fit under
the ESP so that a programming board could be made to use in sandwich
between the ESP and any other board, with only the programming connector
being available.

The programming is made using the GPIO and EN pin so that GPIO16 can
still be used with RESET to keep the watchdog operational.